fix(core-blocks): CB-218, CB-1321 remove html-sanitizer and old components
