--- webapp/packages/core-blocks/src/FormControls/Checkboxes/FieldCheckboxNew.tsx
@@ -1,61 +1,0 @@
-/*
- * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2021 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0.
- * you may not use this file except in compliance with the License.
- */
-
-import styled, { css } from 'reshadow';
-
-import { useStyles } from '@cloudbeaver/core-theming';
-
-import { baseFormControlStylesNew } from '../baseFormControlStylesNew';
-import { isControlPresented } from '../isControlPresented';
-import { Checkbox, CheckboxBaseProps, CheckboxType, ICheckboxControlledProps, ICheckboxObjectProps } from './Checkbox';
-
-const fieldCheckboxStyles = css`
-  Checkbox {
-    margin: -10px;
-  }
-  field {
-    display: flex;
-    align-items: flex-end;
-    white-space: pre-wrap;
-  }
-  field-label {
-    composes: theme-typography--body2 from global;
-    cursor: pointer;
-    user-select: none;
-    padding-left: 10px;
-    line-height: 16px;
-  }
-  Checkbox[disabled] + field-label {
-    cursor: auto;
-  }
-`;
-
-export const FieldCheckboxNew: CheckboxType = function FieldCheckboxNew({
-  children,
-  className,
-  ...rest
-}: CheckboxBaseProps & (ICheckboxControlledProps | ICheckboxObjectProps<any>)) {
-  const styles = useStyles(baseFormControlStylesNew, fieldCheckboxStyles);
-
-  if (rest.autoHide && !isControlPresented(rest.name, rest.state)) {
-    return null;
-  }
-
-  return styled(styles)(
-    <field className={className}>
-      <Checkbox {...(rest as CheckboxBaseProps & ICheckboxControlledProps)} />
-      <field-label
-        htmlFor={rest.id || rest.name}
-        title={rest.title}
-        as="label"
-      >
-        {children}
-      </field-label>
-    </field>
-  );
-};
