+++ webapp/packages/core-blocks/src/FormControls/FormFieldDescription.tsx
@@ -6,32 +6,37 @@
  * you may not use this file except in compliance with the License.
  */
 
-import styled, { use } from 'reshadow';
+import styled, { css } from 'reshadow';
 
 import { useStyles } from '@cloudbeaver/core-theming';
 
 import { baseFormControlStyles } from './baseFormControlStyles';
 
+const style = css`
+  field-label {
+    composes: theme-typography--body1 from global;
+    font-weight: 500;
+  }
+`;
+
 interface Props {
   label?: string;
-  long?: boolean;
-  short?: boolean;
-  raw?: boolean;
+  title?: string;
   className?: string;
 }
 
 export const FormFieldDescription: React.FC<Props> = function FormFieldDescription({
   label,
-  long,
-  short,
-  raw,
+  title,
   children,
   className,
 }) {
-  return styled(useStyles(baseFormControlStyles))(
-    <field as='div' className={className} {...use({ long, short, raw })}>
+  const styles = useStyles(baseFormControlStyles, style);
+
+  return styled(styles)(
+    <field title={title} className={className}>
       {label && <field-label as='label'>{label}</field-label>}
-      <field-description as='div'>
+      <field-description>
         {children}
       </field-description>
     </field>
