+++ webapp/packages/core-blocks/src/ObjectPropertyInfo/ObjectPropertyInfoForm/ObjectPropertyInfoForm.tsx
@@ -8,127 +8,17 @@
 
 import { observer } from 'mobx-react-lite';
 import { useCallback } from 'react';
-import styled from 'reshadow';
 
-import { InputField } from '@cloudbeaver/core-blocks';
 import type { ObjectPropertyInfo } from '@cloudbeaver/core-sdk';
-import { useStyles } from '@cloudbeaver/core-theming';
 
-import { FieldCheckbox } from '../../FormControls/Checkboxes/FieldCheckbox';
-import { Combobox } from '../../FormControls/Combobox';
-import { FormFieldDescription } from '../../FormControls/FormFieldDescription';
-import { FormGroup } from '../../FormControls/FormGroup';
-import { isControlPresented } from '../../FormControls/isControlPresented';
-import { Link } from '../../Link';
+import type { ILayoutSizeProps } from '../../Containers/ILayoutSizeProps';
 import { TextPlaceholder } from '../../TextPlaceholder';
-import { formStyles } from './formStyles';
+import { RenderField } from './RenderField';
 
-const RESERVED_KEYWORDS = ['no', 'off', 'new-password'];
-
-interface RenderFieldProps {
-  property: ObjectPropertyInfo;
-  state: Record<string, any>;
-  editable?: boolean;
-  autofillToken?: string;
-  disabled?: boolean;
-  readOnly?: boolean;
-  autoHide?: boolean;
-  showRememberTip?: boolean;
-  onFocus?: (event: React.FocusEvent<HTMLInputElement>) => void;
-}
-
-const RenderField = observer<RenderFieldProps>(function RenderField({
-  property,
-  state,
-  editable = true,
-  autofillToken = '',
-  disabled,
-  readOnly,
-  autoHide,
-  showRememberTip,
-  onFocus,
-}) {
-  const href = property.features.includes('href');
-  const password = property.features.includes('password');
-  const checkbox = property.dataType === 'Boolean';
-  const combobox = property.validValues && property.validValues.length > 0;
-  let description: string | undefined;
-
-  if (href) {
-    return (
-      <FormFieldDescription label={property.displayName} raw>
-        <Link href={state[property.id!]} target='_blank' rel='noopener noreferrer'>{property.description}</Link>
-      </FormFieldDescription>
-    );
-  }
-
-  if (!editable) {
-    if (autoHide && !isControlPresented(property.id!, state)) {
-      return null;
-    }
-    return (
-      <FormFieldDescription label={property.displayName} raw>
-        {state[property.id!]}
-      </FormFieldDescription>
-    );
-  }
-
-  if (showRememberTip && password && property.value) {
-    description = 'Password saved';
-  }
-
-  if (checkbox) {
-    return (
-      <FieldCheckbox
-        id={property.id}
-        name={property.id!}
-        state={state}
-        label={property.displayName}
-        title={property.description}
-        disabled={disabled}
-      />
-    );
-  }
-
-  if (combobox) {
-    return (
-      <Combobox
-        name={property.id!}
-        state={state}
-        items={property.validValues!}
-        keySelector={value => value}
-        valueSelector={value => value}
-        defaultValue={property.defaultValue}
-        mod="surface"
-        title={property.description}
-        disabled={disabled}
-      >
-        {property.displayName ?? ''}
-      </Combobox>
-    );
-  }
-
-  return (
-    <InputField
-      type={password ? 'password' : 'text'}
-      name={property.id!}
-      state={state}
-      description={description}
-      disabled={disabled}
-      readOnly={readOnly}
-      autoHide={autoHide}
-      autoComplete={RESERVED_KEYWORDS.includes(autofillToken) ? autofillToken : `${autofillToken} ${property.id}`}
-      mod='surface'
-      onFocus={onFocus}
-    >
-      {property.displayName}
-    </InputField>
-  );
-});
-
-interface ObjectPropertyFormProps {
-  properties: ObjectPropertyInfo[] | undefined;
-  state: Record<string, string | number>;
+interface ObjectPropertyFormProps extends ILayoutSizeProps {
+  properties: ObjectPropertyInfo[];
+  state?: Record<string, any>;
+  category?: string | null;
   editable?: boolean;
   autofillToken?: string;
   className?: string;
@@ -142,32 +32,36 @@ interface ObjectPropertyFormProps {
 export const ObjectPropertyInfoForm = observer<ObjectPropertyFormProps>(function ObjectPropertyInfoForm({
   properties,
   state,
+  category,
   editable = true,
-  autofillToken = '',
   className,
+  autofillToken = '',
   disabled,
   readOnly,
   autoHide,
   showRememberTip,
   onFocus,
 }) {
-  const style = useStyles(formStyles);
-
   const handleFocus = useCallback((e: React.FocusEvent<HTMLInputElement>) => {
     if (onFocus) {
       onFocus(e.target.name);
     }
   }, [onFocus]);
 
-  if (!properties || properties.length === 0) {
+  if (properties.length === 0) {
     return <TextPlaceholder>Properties empty</TextPlaceholder>;
   }
 
-  return styled(style)(
-    <form-body as='div' className={className}>
-      {properties.map(property => (
-        <FormGroup key={property.id}>
+  return (
+    <>
+      {properties.map(property => {
+        if (category !== undefined && property.category !== category) {
+          return null;
+        }
+        return (
           <RenderField
+            key={property.id}
+            className={className}
             property={property}
             state={state}
             editable={editable}
@@ -178,8 +72,8 @@ export const ObjectPropertyInfoForm = observer<ObjectPropertyFormProps>(function
             showRememberTip={showRememberTip}
             onFocus={handleFocus}
           />
-        </FormGroup>
-      ))}
-    </form-body>
+        );
+      })}
+    </>
   );
 });
