--- webapp/packages/core-blocks/src/ObjectPropertyInfo/ObjectPropertyInfoForm/ObjectPropertyInfoFormNew.tsx
@@ -1,79 +1,0 @@
-/*
- * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2021 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0.
- * you may not use this file except in compliance with the License.
- */
-
-import { observer } from 'mobx-react-lite';
-import { useCallback } from 'react';
-
-import type { ObjectPropertyInfo } from '@cloudbeaver/core-sdk';
-
-import type { ILayoutSizeProps } from '../../Containers/ILayoutSizeProps';
-import { TextPlaceholder } from '../../TextPlaceholder';
-import { RenderField } from './RenderField';
-
-interface ObjectPropertyFormProps extends ILayoutSizeProps {
-  properties: ObjectPropertyInfo[];
-  state?: Record<string, any>;
-  category?: string | null;
-  editable?: boolean;
-  autofillToken?: string;
-  className?: string;
-  disabled?: boolean;
-  readOnly?: boolean;
-  autoHide?: boolean;
-  showRememberTip?: boolean;
-  onFocus?: (name: string) => void;
-}
-
-export const ObjectPropertyInfoFormNew = observer<ObjectPropertyFormProps>(function ObjectPropertyInfoFormNew({
-  properties,
-  state,
-  category,
-  editable = true,
-  className,
-  autofillToken = '',
-  disabled,
-  readOnly,
-  autoHide,
-  showRememberTip,
-  onFocus,
-}) {
-  const handleFocus = useCallback((e: React.FocusEvent<HTMLInputElement>) => {
-    if (onFocus) {
-      onFocus(e.target.name);
-    }
-  }, [onFocus]);
-
-  if (properties.length === 0) {
-    return <TextPlaceholder>Properties empty</TextPlaceholder>;
-  }
-
-  return (
-    <>
-      {properties.map(property => {
-        if (category !== undefined && property.category !== category) {
-          return null;
-        }
-        return (
-          <RenderField
-            key={property.id}
-            className={className}
-            property={property}
-            state={state}
-            editable={editable}
-            autofillToken={autofillToken}
-            disabled={disabled}
-            readOnly={readOnly}
-            autoHide={autoHide}
-            showRememberTip={showRememberTip}
-            onFocus={handleFocus}
-          />
-        );
-      })}
-    </>
-  );
-});
