--- webapp/packages/core-blocks/src/ObjectPropertyInfo/ObjectPropertyInfoForm/formStyles.ts
@@ -1,18 +1,0 @@
-/*
- * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2021 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0.
- * you may not use this file except in compliance with the License.
- */
-
-import { css } from 'reshadow';
-
-export const formStyles = css`
-  form-body {
-    flex: 1;
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: column;
-  }
-`;
