--- webapp/packages/core-blocks/src/SanitizedHTML/react-sanitized-html.d.ts
@@ -1,14 +1,0 @@
-declare module 'react-sanitized-html' {
-
-  import { Component } from 'react';
-
-  interface ReactSanitizedHtmlProps {
-    html: string;
-    // see https://github.com/apostrophecms/sanitize-html documentation
-    allowedTags?: string[];
-    nonTextTags?: string[];
-  }
-
-  export default class ReactSanitizedHtml extends Component<ReactSanitizedHtmlProps, any> {
-  }
-}
