+++ webapp/packages/core-blocks/src/index.ts
@@ -16,7 +16,6 @@ export * from './layout/AppLogo';
 export * from './layout/TopAppBar';
 export * from './Loader/Loader';
 export * from './ObjectPropertyInfo/ObjectPropertyInfoForm/ObjectPropertyInfoForm';
-export * from './ObjectPropertyInfo/ObjectPropertyInfoForm/ObjectPropertyInfoFormNew';
 export * from './ObjectPropertyInfo/useObjectPropertyCategories';
 
 export * from './Placeholder/Placeholder';
@@ -25,7 +24,6 @@ export * from './Placeholder/PlaceholderContainer';
 export * from './PropertiesTable/PropertiesTable';
 export * from './PropertiesTable/IProperty';
 
-export * from './SanitizedHTML/SanitizedHTML';
 export * from './Slide/SlideBox';
 export * from './Slide/SlideBoxStyles';
 export * from './Slide/SlideElement';
@@ -98,10 +96,8 @@ export * from './ToolsPanel/ToolsAction';
 export * from './ToolsPanel/ToolsPanel';
 export * from './FormControls/Checkboxes/Checkbox';
 export * from './FormControls/Checkboxes/FieldCheckbox';
-export * from './FormControls/Checkboxes/FieldCheckboxNew';
 export * from './FormControls/Checkboxes/CheckboxMarkup';
 export * from './FormControls/Checkboxes/Switch';
-export * from './FormControls/Checkboxes/SwitchNew';
 export * from './FormControls/Checkboxes/useCheckboxState';
 export * from './FormControls/FormBox';
 export * from './FormControls/Filter';
@@ -117,21 +113,16 @@ export * from './Containers/IContainerProps';
 export * from './Containers/ILayoutSizeProps';
 export * from './FormControls/FormBoxElement';
 export * from './FormControls/Combobox';
-export * from './FormControls/ComboboxNew';
 export * from './FormControls/FormContext';
 export * from './FormControls/FormFieldDescription';
-export * from './FormControls/FormFieldDescriptionNew';
 export * from './FormControls/FormGroup';
 export * from './FormControls/InputField';
-export * from './FormControls/InputFieldNew';
-export * from './FormControls/InputGroup';
 export * from './FormControls/Radio';
 export * from './FormControls/RadioGroup';
 export * from './FormControls/RadioGroupContext';
 export * from './FormControls/ShadowInput';
 export * from './FormControls/SubmittingForm';
 export * from './FormControls/Textarea';
-export * from './FormControls/TextareaNew';
 export * from './Link';
 export * from './Cell';
 export * from './ErrorMessage';
@@ -143,6 +134,7 @@ export * from './IconOrImage';
 export * from './StaticImage';
 export * from './TextPlaceholder';
 export * from './InfoItem';
+export * from './Iframe';
 export * from './useClickEvents';
 export * from './useClipboard';
 export * from './useDataResource';
