+++ webapp/packages/core-connections/src/ConnectionForm/Options/ParametersForm.tsx
@@ -7,9 +7,9 @@
  */
 
 import { observer } from 'mobx-react-lite';
-import styled, { css, use } from 'reshadow';
+import styled from 'reshadow';
 
-import { FormGroup, InputField } from '@cloudbeaver/core-blocks';
+import { BASE_CONTAINERS_STYLES, Container, InputField } from '@cloudbeaver/core-blocks';
 import { useTranslate } from '@cloudbeaver/core-localization';
 import type { ConnectionConfig } from '@cloudbeaver/core-sdk';
 import { useStyles } from '@cloudbeaver/core-theming';
@@ -19,66 +19,54 @@ interface Props {
   disabled?: boolean;
   embedded?: boolean;
   readOnly?: boolean;
+  originLocal?: boolean;
 }
-
-const parametersFormStyles = css`
-  sub-label {
-    composes: theme-typography--caption from global;
-    line-height: 14px;
-  }`;
 
 export const ParametersForm = observer<Props>(function ParametersForm({
   config,
   embedded,
   disabled,
   readOnly,
+  originLocal,
 }) {
   const translate = useTranslate();
 
-  return styled(useStyles(parametersFormStyles))(
-    <>
+  return styled(useStyles(BASE_CONTAINERS_STYLES))(
+    <Container gap>
       {!embedded && (
-        <layout-grid-inner as="div">
-          <layout-grid-cell as='div' {...use({ 'span-tablet': 12, 'span-desktop': 8 })}>
-            <InputField
-              type="text"
-              name="host"
-              state={config}
-              disabled={disabled}
-              readOnly={readOnly}
-              mod='surface'
-            >
-              {translate('customConnection_custom_host')}
-              <sub-label as="div">{translate('customConnection_custom_obligatory')}</sub-label>
-            </InputField>
-          </layout-grid-cell>
-          <layout-grid-cell as='div' {...use({ 'span-tablet': 12, 'span-desktop': 4 })}>
-            <InputField
-              type="number"
-              name="port"
-              state={config}
-              disabled={disabled}
-              readOnly={readOnly}
-              mod='surface'
-              short
-            >
-              {translate('customConnection_custom_port')}
-            </InputField>
-          </layout-grid-cell>
-        </layout-grid-inner>
+        <Container wrap gap>
+          <InputField
+            type="text"
+            name="host"
+            state={config}
+            disabled={disabled}
+            readOnly={readOnly || !originLocal}
+            small
+            required
+          >
+            {translate('customConnection_custom_host')}
+          </InputField>
+          <InputField
+            type="number"
+            name="port"
+            state={config}
+            disabled={disabled}
+            readOnly={readOnly || !originLocal}
+            tiny
+          >
+            {translate('customConnection_custom_port')}
+          </InputField>
+        </Container>
       )}
-      <FormGroup>
-        <InputField
-          type="text"
-          name="databaseName"
-          state={config}
-          disabled={disabled}
-          readOnly={readOnly}
-          mod='surface'
-        >
-          {translate('customConnection_custom_database')}
-        </InputField>
-      </FormGroup>
-    </>
+      <InputField
+        type="text"
+        name="databaseName"
+        state={config}
+        disabled={disabled}
+        readOnly={readOnly}
+      >
+        {translate('customConnection_custom_database')}
+      </InputField>
+    </Container>
   );
 });
