+++ webapp/packages/core-connections/src/ConnectionForm/OriginInfo/OriginInfo.tsx
@@ -10,7 +10,7 @@ import { observer } from 'mobx-react-lite';
 import styled, { css } from 'reshadow';
 
 import { UserInfoResource } from '@cloudbeaver/core-authentication';
-import { TextPlaceholder, useTab, Loader, useTabState, ExceptionMessage, useMapResource, ColoredContainer, Group, ObjectPropertyInfoFormNew, BASE_CONTAINERS_STYLES } from '@cloudbeaver/core-blocks';
+import { TextPlaceholder, useTab, Loader, useTabState, ExceptionMessage, useMapResource, ColoredContainer, Group, ObjectPropertyInfoForm, BASE_CONTAINERS_STYLES } from '@cloudbeaver/core-blocks';
 import type { TabContainerPanelComponent } from '@cloudbeaver/core-blocks';
 import { useService } from '@cloudbeaver/core-di';
 import { useTranslate } from '@cloudbeaver/core-localization';
@@ -103,7 +103,7 @@ export const OriginInfo: TabContainerPanelComponent<IConnectionFormProps> = obse
   return styled(styles)(
     <ColoredContainer parent>
       <Group large gap>
-        <ObjectPropertyInfoFormNew
+        <ObjectPropertyInfoForm
           properties={connection.data.origin.details}
           state={state}
           readOnly
